adfing makefile awith some  organisation
